pyproject.toml
[tool.poetry]
name = "practice 1"
version = "0.1.0"
description = ""
authors = ["Daniyal <daniyal.mambetov.04@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.1.5"
djangorestframework = "^3.15.2"
django-filter = "^24.3"
django-rest-framework = "^0.1.0"
djangorestframework-simplejwt = "^5.4.0"
django-bootstrap = "^0.2.4"
django-cors-headers = "^4.6.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
